Web: Files: refactoring: created withContextOptions HOC, refactoring withFileActions
